test: Section 클래스 인수 테스트 추가 및 이동
