+++ src/main/java/nextstep/subway/dto/StationResponse.java
@@ -47,4 +47,12 @@ public class StationResponse {
     public int hashCode() {
         return Objects.hash(name);
     }
+
+    @Override
+    public String toString() {
+        return "StationResponse{" +
+                "id=" + id +
+                ", name='" + name + '\'' +
+                '}';
+    }
 }
