+++ src/test/java/nextstep/subway/station/LineAcceptanceTest.java
@@ -190,154 +190,6 @@ public class LineAcceptanceTest {
     }
 
     /**
-     * Given 노선을 생성하고
-     * When 구간을 추가하면
-     * Then 노선에 속한 지하철역의 숫자가 증가한다
-     */
-    @DisplayName("역 사이에 새로운 역을 등록할 경우")
-    @Test
-    void addSectionCase1() {
-        // given
-        LineResponse line = 응답_객체_생성(노선_등록("2호선", "초록", 7, A역.getId(), C역.getId()), LineResponse.class);
-
-        // when
-        구간_추가(A역.getId(), B역.getId(), 4, line.getId());
-
-        // then
-        LineResponse find = 응답_객체_생성(노선_조회(line.getId()), LineResponse.class);
-        assertThat(find.getStations()).hasSize(3);
-    }
-
-    @DisplayName("역 사이에 새로운 역을 등록할 경우")
-    @Test
-    void addSectionCase2() {
-        // given
-        LineResponse line = 응답_객체_생성(노선_등록("2호선", "초록", 7, A역.getId(), C역.getId()), LineResponse.class);
-
-        // when
-        구간_추가(B역.getId(), C역.getId(), 4, line.getId());
-
-        // then
-        LineResponse find = 응답_객체_생성(노선_조회(line.getId()), LineResponse.class);
-        assertThat(find.getStations()).hasSize(3);
-    }
-
-    /**
-     * Given 노선을 생성하고
-     * When 구간을 추가하면
-     * Then 노선에 속한 지하철역의 숫자가 증가한다
-     */
-    @DisplayName("새로운 역을 상행 종점으로 등록할 경우")
-    @Test
-    void addSectionCase3() {
-        // given
-        LineResponse line = 응답_객체_생성(노선_등록("2호선", "초록", 7, A역.getId(), C역.getId()), LineResponse.class);
-
-        // when
-        구간_추가(B역.getId(), A역.getId(), 4, line.getId());
-
-        // then
-        LineResponse find = 응답_객체_생성(노선_조회(line.getId()), LineResponse.class);
-        assertThat(find.getStations()).hasSize(3);
-    }
-
-    /**
-     * Given 노선을 생성하고
-     * When 구간을 추가하면
-     * Then 노선에 속한 지하철역의 숫자가 증가한다
-     */
-    @DisplayName("새로운 역을 하행 종점으로 등록할 경우")
-    @Test
-    void addSectionCase4() {
-        // given
-        LineResponse line = 응답_객체_생성(노선_등록("2호선", "초록", 7, A역.getId(), C역.getId()), LineResponse.class);
-
-        // when
-        구간_추가(C역.getId(), B역.getId(), 3, line.getId());
-
-        // then
-        LineResponse find = 응답_객체_생성(노선_조회(line.getId()), LineResponse.class);
-        assertThat(find.getStations()).hasSize(3);
-    }
-
-    /**
-     * Given 노선을 생성하고
-     * When 구간을 추가하면
-     * Then 노선에 속한 지하철역의 숫자가 증가한다
-     */
-    @DisplayName("다양한 패턴으로 노선에 지하철역 등록할 경우")
-    @Test
-    void addSectionCase5() {
-        LineResponse line = 응답_객체_생성(노선_등록("2호선", "초록", 15, A역.getId(), D역.getId()), LineResponse.class);
-
-        구간_추가(A역.getId(), B역.getId(), 4, line.getId());
-        구간_추가(C역.getId(), D역.getId(), 3, line.getId());
-        구간_추가(E역.getId(), A역.getId(), 15, line.getId());
-        구간_추가(D역.getId(), F역.getId(), 30, line.getId());
-
-        // then
-        LineResponse find = 응답_객체_생성(노선_조회(line.getId()), LineResponse.class);
-        assertThat(find.getStations()).hasSize(6);
-    }
-
-    /**
-     * Given 노선을 생성하고
-     * When 노선 거리를 초과하는 구간을 중간에 추가하면
-     * Then 노선을 추가하지 못한다
-     */
-    @DisplayName("구간 추가 오류(구간 거리 초과1)")
-    @Test
-    void addSectionException1() {
-        // given
-        LineResponse line = 응답_객체_생성(노선_등록("2호선", "초록", 7, A역.getId(), C역.getId()), LineResponse.class);
-
-        // when
-        ValidatableResponse response = 구간_추가(A역.getId(), B역.getId(), 7, line.getId());
-
-        // then
-        응답_검증(response, HttpStatus.BAD_REQUEST);
-    }
-
-    @DisplayName("구간 추가 오류(구간 거리 초과2)")
-    @Test
-    void addSectionException2() {
-        // given
-        LineResponse line = 응답_객체_생성(노선_등록("2호선", "초록", 7, A역.getId(), C역.getId()), LineResponse.class);
-
-        // when
-        ValidatableResponse response = 구간_추가(B역.getId(), C역.getId(), 8, line.getId());
-
-        // then
-        응답_검증(response, HttpStatus.BAD_REQUEST);
-    }
-
-    @DisplayName("구간 추가 오류(노선에 이미 포함된 지하철 역)")
-    @Test
-    void addSectionException3() {
-        // given
-        LineResponse line = 응답_객체_생성(노선_등록("2호선", "초록", 7, A역.getId(), C역.getId()), LineResponse.class);
-
-        // when
-        ValidatableResponse response = 구간_추가(A역.getId(), B역.getId(), 7, line.getId());
-
-        // then
-        응답_검증(response, HttpStatus.BAD_REQUEST);
-    }
-
-    @DisplayName("구간 추가 오류(노선에 모두 포함안된 지하철 역)")
-    @Test
-    void addSectionException4() {
-        // given
-        LineResponse line = 응답_객체_생성(노선_등록("2호선", "초록", 7, A역.getId(), C역.getId()), LineResponse.class);
-
-        // when
-        ValidatableResponse response = 구간_추가(E역.getId(), F역.getId(), 10, line.getId());
-
-        // then
-        응답_검증(response, HttpStatus.BAD_REQUEST);
-    }
-
-    /**
      * When 존재하지 않는 지하철역으로 노선을 등록하면
      * Then 노선 등록이 불가하다.
      */
@@ -378,16 +230,6 @@ public class LineAcceptanceTest {
                 .body(lineUpdateRequest)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when().put("/lines/" + id)
-                .then().log().all();
-    }
-
-    public static ValidatableResponse 구간_추가(long upStationId, long downStationId, int distance, long lineId) {
-        SectionRequest sectionRequest = new SectionRequest(upStationId, downStationId, distance);
-
-        return RestAssured.given().log().all()
-                .body(sectionRequest)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().post("/lines/" + lineId + "/sections")
                 .then().log().all();
     }
 }
