+++ src/test/java/nextstep/subway/station/SectionAcceptanceTest.java
@@ -3,6 +3,8 @@ package nextstep.subway.station;
 import io.restassured.RestAssured;
 import io.restassured.response.ValidatableResponse;
 import nextstep.subway.dto.LineResponse;
+import nextstep.subway.dto.SectionRequest;
+import nextstep.subway.dto.SectionResponse;
 import nextstep.subway.dto.StationResponse;
 import nextstep.subway.utils.DatabaseCleanup;
 import org.junit.jupiter.api.BeforeEach;
@@ -11,12 +13,15 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 
-import static nextstep.subway.station.LineAcceptanceTest.구간_추가;
+import java.util.List;
+
 import static nextstep.subway.station.LineAcceptanceTest.노선_등록;
-import static nextstep.subway.station.StationAcceptanceTest.응답_객체_생성;
-import static nextstep.subway.station.StationAcceptanceTest.지하철역_등록;
+import static nextstep.subway.station.LineAcceptanceTest.노선_조회;
+import static nextstep.subway.station.StationAcceptanceTest.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("구간 관련 기능")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
@@ -51,6 +56,154 @@ public class SectionAcceptanceTest {
     }
 
     /**
+     * Given 노선을 생성하고
+     * When 구간을 추가하면
+     * Then 노선에 속한 지하철역의 숫자가 증가한다
+     */
+    @DisplayName("역 사이에 새로운 역을 등록할 경우")
+    @Test
+    void addSectionCase1() {
+        // given
+        LineResponse line = 응답_객체_생성(노선_등록("2호선", "초록", 7, A역.getId(), C역.getId()), LineResponse.class);
+
+        // when
+        구간_추가(A역.getId(), B역.getId(), 4, line.getId());
+
+        // then
+        LineResponse find = 응답_객체_생성(노선_조회(line.getId()), LineResponse.class);
+        assertThat(find.getStations()).hasSize(3);
+    }
+
+    @DisplayName("역 사이에 새로운 역을 등록할 경우")
+    @Test
+    void addSectionCase2() {
+        // given
+        LineResponse line = 응답_객체_생성(노선_등록("2호선", "초록", 7, A역.getId(), C역.getId()), LineResponse.class);
+
+        // when
+        구간_추가(B역.getId(), C역.getId(), 4, line.getId());
+
+        // then
+        LineResponse find = 응답_객체_생성(노선_조회(line.getId()), LineResponse.class);
+        assertThat(find.getStations()).hasSize(3);
+    }
+
+    /**
+     * Given 노선을 생성하고
+     * When 구간을 추가하면
+     * Then 노선에 속한 지하철역의 숫자가 증가한다
+     */
+    @DisplayName("새로운 역을 상행 종점으로 등록할 경우")
+    @Test
+    void addSectionCase3() {
+        // given
+        LineResponse line = 응답_객체_생성(노선_등록("2호선", "초록", 7, A역.getId(), C역.getId()), LineResponse.class);
+
+        // when
+        구간_추가(B역.getId(), A역.getId(), 4, line.getId());
+
+        // then
+        LineResponse find = 응답_객체_생성(노선_조회(line.getId()), LineResponse.class);
+        assertThat(find.getStations()).hasSize(3);
+    }
+
+    /**
+     * Given 노선을 생성하고
+     * When 구간을 추가하면
+     * Then 노선에 속한 지하철역의 숫자가 증가한다
+     */
+    @DisplayName("새로운 역을 하행 종점으로 등록할 경우")
+    @Test
+    void addSectionCase4() {
+        // given
+        LineResponse line = 응답_객체_생성(노선_등록("2호선", "초록", 7, A역.getId(), C역.getId()), LineResponse.class);
+
+        // when
+        구간_추가(C역.getId(), B역.getId(), 3, line.getId());
+
+        // then
+        LineResponse find = 응답_객체_생성(노선_조회(line.getId()), LineResponse.class);
+        assertThat(find.getStations()).hasSize(3);
+    }
+
+    /**
+     * Given 노선을 생성하고
+     * When 구간을 추가하면
+     * Then 노선에 속한 지하철역의 숫자가 증가한다
+     */
+    @DisplayName("다양한 패턴으로 노선에 지하철역 등록할 경우")
+    @Test
+    void addSectionCase5() {
+        LineResponse line = 응답_객체_생성(노선_등록("2호선", "초록", 15, A역.getId(), D역.getId()), LineResponse.class);
+
+        구간_추가(A역.getId(), B역.getId(), 4, line.getId());
+        구간_추가(C역.getId(), D역.getId(), 3, line.getId());
+        구간_추가(E역.getId(), A역.getId(), 15, line.getId());
+        구간_추가(D역.getId(), F역.getId(), 30, line.getId());
+
+        // then
+        LineResponse find = 응답_객체_생성(노선_조회(line.getId()), LineResponse.class);
+        assertThat(find.getStations()).hasSize(6);
+    }
+
+    /**
+     * Given 노선을 생성하고
+     * When 노선 거리를 초과하는 구간을 중간에 추가하면
+     * Then 노선을 추가하지 못한다
+     */
+    @DisplayName("구간 추가 오류(구간 거리 초과1)")
+    @Test
+    void addSectionException1() {
+        // given
+        LineResponse line = 응답_객체_생성(노선_등록("2호선", "초록", 7, A역.getId(), C역.getId()), LineResponse.class);
+
+        // when
+        ValidatableResponse response = 구간_추가(A역.getId(), B역.getId(), 7, line.getId());
+
+        // then
+        응답_검증(response, HttpStatus.BAD_REQUEST);
+    }
+
+    @DisplayName("구간 추가 오류(구간 거리 초과2)")
+    @Test
+    void addSectionException2() {
+        // given
+        LineResponse line = 응답_객체_생성(노선_등록("2호선", "초록", 7, A역.getId(), C역.getId()), LineResponse.class);
+
+        // when
+        ValidatableResponse response = 구간_추가(B역.getId(), C역.getId(), 8, line.getId());
+
+        // then
+        응답_검증(response, HttpStatus.BAD_REQUEST);
+    }
+
+    @DisplayName("구간 추가 오류(노선에 이미 포함된 지하철 역)")
+    @Test
+    void addSectionException3() {
+        // given
+        LineResponse line = 응답_객체_생성(노선_등록("2호선", "초록", 7, A역.getId(), C역.getId()), LineResponse.class);
+
+        // when
+        ValidatableResponse response = 구간_추가(A역.getId(), B역.getId(), 7, line.getId());
+
+        // then
+        응답_검증(response, HttpStatus.BAD_REQUEST);
+    }
+
+    @DisplayName("구간 추가 오류(노선에 모두 포함안된 지하철 역)")
+    @Test
+    void addSectionException4() {
+        // given
+        LineResponse line = 응답_객체_생성(노선_등록("2호선", "초록", 7, A역.getId(), C역.getId()), LineResponse.class);
+
+        // when
+        ValidatableResponse response = 구간_추가(E역.getId(), F역.getId(), 10, line.getId());
+
+        // then
+        응답_검증(response, HttpStatus.BAD_REQUEST);
+    }
+
+    /**
      * Given 노선 및 구간을 설정한다.
      * When, then 노선 내 순서대로 구간을 조회한다.
      */
@@ -65,13 +218,24 @@ public class SectionAcceptanceTest {
         구간_추가(D역.getId(), F역.getId(), 30, line.getId());
 
         // when, then
-        노선_구간_조회(line.getId());
+        List<SectionResponse> responses = 구간_목록_조회(line.getId());
+        assertThat(responses).hasSize(5);
     }
 
-    public static ValidatableResponse 노선_구간_조회(Long lineId) {
+    public static ValidatableResponse 구간_추가(long upStationId, long downStationId, int distance, long lineId) {
+        SectionRequest sectionRequest = new SectionRequest(upStationId, downStationId, distance);
+
         return RestAssured.given().log().all()
+                .body(sectionRequest)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when().post("/lines/" + lineId + "/sections")
+                .then().log().all();
+    }
+
+    public static List<SectionResponse> 구간_목록_조회(Long lineId) {
+        ValidatableResponse listResponse = RestAssured.given().log().all()
                 .when().get("/sections/" + lineId)
                 .then().log().all();
+        return getJsonPathForResponse(listResponse).getList("$", SectionResponse.class);
     }
 }
